requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2
PyYAML>=6
requests>=2.31
tenacity>=8
httpx>=0.27
google-cloud-storage>=2.10
coinbase-advanced-py>=1.3.0
python-dotenv>=1.0.1
coinbase-advanced-py>=1.3.0
python-dotenv>=1.0.1

coinbase-advanced-py

tenacity
